+++ src/plugin/utils.ts
@@ -3,3 +3,18 @@ import {v4 as uuidv4} from 'uuid';
 export function generateID() {
     return uuidv4().substring(0, 8);
 }
+
+export function addHiddenProperty<
+    B extends Record<string | symbol, unknown>,
+    F extends string | symbol,
+    V,
+>(box: B, field: F, value: V) {
+    if (!(field in box)) {
+        Object.defineProperty(box, field, {
+            enumerable: false,
+            value: value,
+        });
+    }
+
+    return box as B & {[P in F]: V};
+}
